@@ -11,9 +11,7 @@ import { TextInput, StyleSheet, Text, View, TouchableOpacity, Alert } from 'reac
 import NotifService from './NotifService';
 import appConfig from './app.json';
 
-type Props = {};
-export default class App extends Component<Props> {
-
+export default class App extends Component {
   constructor(props) {
     super(props);
     this.state = {
