@@ -59,15 +59,15 @@ export default class NotifService {
 
       /* iOS only properties */
       alertAction: 'view', // (optional) default: view
-      category: null, // (optional) default: null
-      userInfo: null, // (optional) default: null (object containing additional notification data)
+      category: '', // (optional) default: empty string
+      userInfo: {}, // (optional) default: {} (using null throws a JSON value '<null>' error)
 
       /* iOS and Android properties */
       title: "Local Notification", // (optional)
       message: "My Notification Message", // (required)
       playSound: false, // (optional) default: true
       soundName: 'default', // (optional) Sound to play when the notification is shown. Value of 'default' plays the default sound. It can be set to a custom sound such as 'android.resource://com.xyz/raw/my_sound'. It will look for the 'my_sound' audio file in 'res/raw' directory and play it. default: 'default' (default sound is played)
-      number: '10', // (optional) Valid 32 bit integer specified as string. default: none (Cannot be zero)
+      number: 10, // (optional) Valid 32 bit integer specified as string. default: none (Cannot be zero)
       actions: '["Yes", "No"]',  // (Android only) See the doc for notification actions to know more
     });
   }
@@ -94,8 +94,8 @@ export default class NotifService {
 
       /* iOS only properties */
       alertAction: 'view', // (optional) default: view
-      category: null, // (optional) default: null
-      userInfo: null, // (optional) default: null (object containing additional notification data)
+      category: '', // (optional) default: empty string
+      userInfo: {}, // (optional) default: {} (using null throws a JSON value '<null>' error)
 
       /* iOS and Android properties */
       title: "Scheduled Notification", // (optional)
