@@ -1,4 +1,6 @@
-/** @format */
+/**
+ * @format
+ */
 
 import {AppRegistry} from 'react-native';
 import App from './App';
